requirements.txt
matplotlib==3.4.2
pandas==1.1.1
scikit_image==0.18.3
numpy==1.19.1
Pillow==8.4.0
scikit_learn==1.0.1
skimage==0.0
